+++ dashboards-observability/common/constants/explorer.ts
@@ -117,3 +117,16 @@ export const AGGREGATION_OPTIONS = [
     label: 'STDDEV_POP',
   },
 ];
+
+// Data table constants
+export const GRID_HEADER_COLUMN_MAX_WIDTH = '150px';
+export const GRID_PAGE_RANGE_DISPLAY = 5;
+export const COLUMN_DEFAULT_MIN_WIDTH = 100;
+export const GRID_PAGE_SIZES = [10, 50, 100];
+export const ROW_DENSITIES = [
+  { icon: 'tableDensityExpanded', height: 55, selected: false },
+  { icon: 'tableDensityNormal', height: 45, selected: false },
+  { icon: 'tableDensityCompact', height: 35, selected: true },
+];
+
+export const HEADER_HEIGHT = 35;
