+++ dashboards-observability/public/components/visualizations/charts/lines/line.tsx
@@ -94,7 +94,7 @@ export const Line = ({ visualizations, layout, config }: any) => {
       };
     });
 
-    var layoutForBarMode = {
+    let layoutForBarMode = {
       barmode: 'group',
     };
     const mergedLayout = {
